+++ tempest_fastapi_sdk/__init__.py
@@ -482,7 +482,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.206.0"
+__version__: str = "0.207.0"
 
 __all__: list[str] = [
     "ACCESS_TOKEN_TYPE",